StatsDiscretizations

StatsDiscretizations.RealLineDiscretizer

StatsDiscretizations.unwrap

StatsDiscretizations.wrap

StatsDiscretizations.RealLineDiscretizer — Type

RealLineDiscretizer{L, R, T, G <:AbstractVector{T}} <: AbstractRealLineDiscretizer{Interval{L, R, T}}

An extended continuous discretizer that partitions the real line into intervals based on a given grid of points. It extends the discretization beyond the grid by including intervals from -∞ to the first grid point and from the last grid point to +∞.

The type parameters L and R specify the endpoint types of the intervals (e.g., Closed or Open), T is the element type of the grid, and G is the type of the grid vector.

Fields

grid::G: The grid of points used for discretization.

Examples

grid = [0.0, 1.0, 2.0]
discr = RealLineDiscretizer{Closed,Open}(grid)

StatsDiscretizations.unwrap — Method

unwrap(x)

Unwraps the value x. By default, returns x as is.

StatsDiscretizations.wrap — Method

wrap(x, val)

Wraps the value val based on x. By default, returns val as is.